refactor(integrators): move get_substring_geometry to vector_data_importer/utils
- Add get_substring_geometry as module-level function in utils.py
- Remove static method from LinearIntegrator
- Update all call sites (LineStructurePlacement, substring_feature)
- Move test_get_substring_geometry to test_utils.py with length assertion added
