+++ tests/vector_data_importer/test_utils.py
@@ -500,10 +500,6 @@ def test_compute_selected_ids_unknown_field_warns_and_returns_candidates():
     assert any(issubclass(w.category, FeaturesImporterWarning) for w in caught)
 
 
-# ---------------------------------------------------------------------------
-# build_feature_mapping
-# ---------------------------------------------------------------------------
-
 
 @pytest.fixture
 def simple_layer():
@@ -557,3 +553,21 @@ def test_build_feature_mapping(config, values, expected_keys):
     for feature in layer.getFeatures():
         if feature["code"] in expected_keys:
             assert result[feature["code"]] == feature
+
+@pytest.mark.parametrize("simplify", [True, False])
+def test_get_substring_geometry(simplify):
+    """get_substring_geometry returns a QgsGeometry; simplify reduces it to 2 vertices."""
+    from threedi_schematisation_editor.vector_data_importer.utils import get_substring_geometry
+
+    line_geom = QgsGeometry.fromPolylineXY(
+        [QgsPointXY(0, 0), QgsPointXY(50, 0), QgsPointXY(100, 0)]
+    )
+    curve = line_geom.constGet()
+    result = get_substring_geometry(curve, 25.0, 75.0, simplify=simplify)
+
+    assert isinstance(result, QgsGeometry)
+    assert result.length() == pytest.approx(50.0)
+    if simplify:
+        assert len(result.asPolyline()) == 2
+    else:
+        assert len(result.asPolyline()) >= 2
